+++ src/project/git/GitCommandGitProject.ts
@@ -1,5 +1,6 @@
 import { exec } from "child-process-promise";
 import * as stringify from "json-stringify-safe";
+import { HandlerContext } from "../../HandlerContext";
 
 import axios from "axios";
 import { isLocalProject, ReleaseFunction } from "../local/LocalProject";
@@ -63,7 +64,7 @@ export class GitCommandGitProject extends NodeFsLocalProject implements GitProje
      * @param {DirectoryManager} directoryManager
      * @return {Promise<GitCommandGitProject>}
      */
-    public static cloned(credentials: ProjectOperationCredentials,
+    public static cloned(context: HandlerContext, credentials: ProjectOperationCredentials,
                          id: RemoteRepoRef,
                          opts: CloneOptions = DefaultCloneOptions,
                          directoryManager: DirectoryManager = DefaultDirectoryManager): Promise<GitProject> {