+++ test/api/generatorEndToEndTest.ts
@@ -69,7 +69,7 @@ describe("generator end to end", () => {
                 return hasFile(GitHubToken, TargetOwner, repoName, "pom.xml")
                     .then(r => {
                         assert(r);
-                        return GitCommandGitProject.cloned({ token: GitHubToken },
+                        return GitCommandGitProject.cloned({} as HandlerContext, { token: GitHubToken },
                             new GitHubRepoRef(TargetOwner, repoName))
                             .then(verifyPermissions);
                     });
@@ -83,7 +83,7 @@ describe("generator end to end", () => {
             deleteOrIgnore(repoName).then(done(err));
         };
 
-        const clonedSeed = GitCommandGitProject.cloned({ token: GitHubToken },
+        const clonedSeed = GitCommandGitProject.cloned({} as HandlerContext, { token: GitHubToken },
             new GitHubRepoRef("atomist-seeds", "spring-rest-seed"));
         const targetRepo = new GitHubRepoRef(TargetOwner, repoName);
 
@@ -96,7 +96,7 @@ describe("generator end to end", () => {
                 return hasFile(GitHubToken, TargetOwner, repoName, "pom.xml")
                     .then(r => {
                         assert(r);
-                        return GitCommandGitProject.cloned({ token: GitHubToken },
+                        return GitCommandGitProject.cloned({} as HandlerContext, { token: GitHubToken },
                             targetRepo)
                             .then(verifyPermissions)
                             .then(() => {
@@ -109,7 +109,7 @@ describe("generator end to end", () => {
     it("should refuse to create a new GitHub repo using existing repo name", function(done) {
         this.retries(5);
 
-        const clonedSeed = GitCommandGitProject.cloned({ token: GitHubToken },
+        const clonedSeed = GitCommandGitProject.cloned({} as HandlerContext, { token: GitHubToken },
             new GitHubRepoRef("atomist-seeds", "spring-rest-seed"));
         const targetRepo = new GitHubRepoRef("atomist-travisorg", "this-repository-exists");