+++ test/project/git/GitProjectTest.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
 import "mocha";
+import { HandlerContext } from "../../../src/HandlerContext";
 
 import * as assert from "power-assert";
 import { tempProject } from "../utils";
@@ -30,7 +31,7 @@ describe("GitProject cloning on filesystem", () => {
 
     const getAClone = (repoName: string = RepoName) => {
         const repositoryThatExists = new GitHubRepoRef(Owner, repoName);
-        return GitCommandGitProject.cloned(Creds, repositoryThatExists);
+        return GitCommandGitProject.cloned({} as HandlerContext, Creds, repositoryThatExists);
     };
 
     it("never returns the same place on the filesystem twice at once", done => {
@@ -157,7 +158,7 @@ ding dong ding
 
     it("check out commit", done => {
         const sha = "590ed8f7a2430d45127ea04cc5bdf736fe698712";
-        GitCommandGitProject.cloned(Creds, new GitHubRepoRef("atomist", "microgrammar", sha))
+        GitCommandGitProject.cloned({} as HandlerContext, Creds, new GitHubRepoRef("atomist", "microgrammar", sha))
             .then(p => {
                 checkProject(p);
                 const gp: GitProject = GitCommandGitProject.fromProject(p, Creds);
@@ -168,7 +169,7 @@ ding dong ding
     }).timeout(5000);
 
     it("clones a project subdirectory", done => {
-        GitCommandGitProject.cloned(Creds, new GitHubRepoRef("pallets", "flask", "master",
+        GitCommandGitProject.cloned({} as HandlerContext, Creds, new GitHubRepoRef("pallets", "flask", "master",
             GitHubDotComBase, "examples/flaskr"))
             .then(gp => {
                 assert(!!gp.findFileSync("flaskr/__init__.py"), "Should be able to find file under subdirectory");