+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [0.5.2]: https://github.com/atomist/automation-client-ts/compare/0.5.0...0.5.2
 
+### Changed
+**Breaking** Add parameter "context: HandlerContext" to GitCommandGitProject.cloned
+**Breaking** Add parameter "context: HandlerContext" to DifferenceEngine.run
+**Breaking** Add parameter "context: HandlerContext" to gitHubRepoLoader
+**Breaking** Add parameter "context: HandlerContext" to defaultRepoLoader
+**Breaking** Add parameter "context: HandlerContext" to generatorHandler
+
 ### Added
 
 -   GitHubUserLogin and AtomistWebhookUrlBase mapped parameter helpers